In essence this merge achieves:

  1. unifying the stdlib with stdlib-follow
  2. moving to the stdlib-follow build system

The merging operation was relatively straightforward, except for the build
system. To not undo any work there, I omitted some configure.ac config. Within
the stdlib-follow build system, the configuration and building is done in four
parts:

  1. call ./bootstrap
    a. detect all source files (.sac, .xsac, .c, .h, etc.)
       and store the location of these in a makefile
    b. call autotools and generate the configure files both in the root
       directory and under sub/.
  2. call configure in the root directory which generates the makefiles
  3. call make in the root directory
    a. analysis detected sources files from before for dependencies: this
       generates a TON of *.moddep, *.objdep, *.donedep, etc files which
       provide the targets for the build system
    b. calls the configure in sub/ to generate the makefiles for both the seq
       and mt_pth targets (more can be added later).
    c. calls the make in sub/ compiling the stdlib for targets
  4. call make install which copies the compiled libraries on the host

Below is a collection of the commits and file changes caused by this merge:

* master: (25 commits)
  Added function to return the elapsed time from the interval
  Added MTClock and time unit selection
  Change FibreIO float/double output format
  Added a new clock facility
  Updated gitignore
  Fix for FibreIO float number parsing
  Committing non-distributed memory backend-variant
  Fixed missing dependency for previous commit
  Fixed warnings
  Added printing mechanism for complex numbers
  Added fprintf/printf for CUDA
  Distributed memory backend: Arrays of all types are supported
  Printing for distributed memory backend, initial version
  Fix fault in shift() - very wrong answers. It now also AWLFs a fair bit
    better!
  Kill dead comment text.
  Fix deprecated (:int).
  Disable custom simd operations for standard library.
  Make glibc 2.20 shut up about deprecated _BSD_SOURCE.
  Make gettime of RTclock match clock_gettime sec and nsec types.
  Fix errors produced by glibc-2.20, except _DEFAULT_SOURCE.
  ...

 Changes to be committed:
	modified:   .gitignore
	modified:   bootstrap
	modified:   build-aux/saclib-deps.mk
	modified:   src/classes/random/Rand.sac
	modified:   src/classes/random/Rand48.sac
	modified:   src/classes/random/Random.sac
	modified:   src/modules/auxiliary/C99Benchmarking.sac
	modified:   src/modules/auxiliary/src/C99Benchmarking/bench.c
	modified:   src/modules/numerical/Math.sac
	modified:   src/modules/structures/ArrayTransform.xsac
	modified:   src/modules/structures/ComplexBasics.sac
	modified:   src/modules/structures/Constants.sac
	new file:   src/modules/structures/src/Constants/epidouble.c
	modified:   src/modules/structures/src/Constants/mindouble.c
	new file:   src/modules/structures/src/Constants/tinydouble.c
	new file:   src/world/stdio/ArrayIO.sac
	modified:   src/world/stdio/ArrayIO.xsac
	modified:   src/world/stdio/ComplexIO.sac
	modified:   src/world/stdio/TermFile.sac
	new file:   src/world/stdio/opencv/CvCapture.sac
	new file:   src/world/stdio/opencv/IplImage.sac
	new file:   src/world/stdio/opencv/src/cvcapture.c
	new file:   src/world/stdio/opencv/src/iplimage.c
	modified:   src/world/stdio/src/ArrayIO/PrintArray.c
	modified:   src/world/stdio/src/BinFile/binfWriteDblArr.c
	new file:   src/world/stdio/src/ComplexIO/Makefile
	new file:   src/world/stdio/src/ComplexIO/PrintComplexArray.c
	new file:   src/world/stdio/src/FibreIO/.gitignore
	modified:   src/world/stdio/src/FibreIO/FibrePrint.c
	modified:   src/world/stdio/src/FibreIO/FibreScan.y
	new file:   src/world/stdio/src/TermFile/printf.c
	modified:   src/world/system/FileSystem.sac
	new file:   src/world/system/MTClock.sac
	modified:   src/world/system/Process.sac
	modified:   src/world/system/RTClock.sac
	modified:   src/world/system/src/FileSystem/FileSystem.h
	new file:   src/world/system/src/FileSystem/mktemp.c
	modified:   src/world/system/src/FileSystem/pltmp.c
	new file:   src/world/system/src/MTClock/Makefile
	new file:   src/world/system/src/MTClock/mtclock.c
	new file:   src/world/system/src/Process/system.c
	modified:   src/world/system/src/RTClock/rtclock.c

+++ src/system/src/RTClock/rtclock.c
@@ -32,7 +32,7 @@ void SAC_RTClock_touch( void *rtclock)
 {
 }
 
-void SAC_RTClock_gettime( int *sec, int *nano)
+void SAC_RTClock_gettime( long *sec, long *nano)
 {
   struct timespec result;
 
@@ -56,6 +56,6 @@ void SAC_RTClock_gettime( int *sec, int *nano)
 #warning no implementation of SAC_RTClock_gettime on this target
 #endif
 
-  *sec = (int) result.tv_sec;
-  *nano = (int) result.tv_nsec;
+  *sec = result.tv_sec;
+  *nano = result.tv_nsec;
 }
